refacto (modules/RawAttachment) move interface from activity to utils
